+++ metro.config.js
@@ -4,17 +4,23 @@ const glob = require('glob');
 const metroResolve = require('metro-resolver/src/resolve');
 
 const domModuleMap = {
+  // delegate to "web"
   './effects/BlurView': path.resolve(__dirname, './src/BlurView.dom.js'),
   './NativeLinearGradient': path.resolve(__dirname, './src/NativeLinearGradient.dom.js'),
-  './NativeModulesProxy': path.resolve(__dirname, './src/NativeModulesProxy.dom.js'),
-  'expo-asset/src/Asset': path.resolve(__dirname, './src/ExpoAsset.dom.js'),
-  'expo-asset': path.resolve(__dirname, './src/ExpoAssetIndex.dom.js'),
   './facebook-ads': path.resolve(__dirname, './src/facebook-ads.dom.js'),
 
+  // custom patches
+  './NativeModulesProxy': path.resolve(__dirname, './src/NativeModulesProxy.dom.js'),
+  './NativeViewManagerAdapter': path.resolve(__dirname, './src/NativeViewManagerAdapter.dom.js'),
+  'expo-asset/src/Asset': path.resolve(__dirname, './src/ExpoAsset.dom.js'),
+  'expo-asset': path.resolve(__dirname, './src/ExpoAssetIndex.dom.js'),
+
+  // vendor libs
+  'expo-camera': path.resolve(__dirname, './vendor/expo-camera/index.ts'),
   'react-native-gesture-handler': path.resolve(__dirname, './vendor/react-native-gesture-handler/index.js'),
 };
 
-// pass mapping workaround for Expo SDK
+// path mapping workaround for Expo SDK
 const baseDir = path.resolve(__dirname, '../react-native-dom/Libraries');
 glob.sync(baseDir + '/**/*.dom.js').forEach((fullPath) => {
   const relPath = fullPath.slice(baseDir.length, -'.dom.js'.length);
@@ -26,7 +32,7 @@ glob.sync(baseDir + '/**/*.dom.js').forEach((fullPath) => {
   domModuleMap['./' + name] = fullPath;
   domModuleMap['./../../' + dir + '/' + name] = fullPath;
 });
-require('fs').appendFileSync('/tmp/metro-resolve.log', `${JSON.stringify(domModuleMap, null, 2)}\n`);
+// require('fs').appendFileSync('/tmp/metro-resolve.log', `${JSON.stringify(domModuleMap)}\n`);
 
 module.exports = {
   projectRoot: path.resolve(__dirname, '../..'),